Rename composed-ui to extensions
