--- libs/composed-ui/angular/banner/README.md
@@ -1,3 +1,0 @@
-# @kirbydesign/composed-ui-angular/banner
-
-Secondary entry point of `@kirbydesign/composed-ui-angular`. It can be used by importing from `@kirbydesign/composed-ui-angular/banner`.
