--- libs/composed-ui/angular/index.ts
@@ -1,1 +1,0 @@
-export * from '@kirbydesign/composed-ui-angular/banner';
