+++ libs/extensions/angular/README.md
@@ -1,0 +1,4 @@
+# Kirby Extensions Angular
+
+This library contains... 
+2. [Contributing](CONTRIBUTING.md)
