+++ libs/extensions/angular/banner/README.md
@@ -1,0 +1,3 @@
+# @kirbydesign/extensions-angular/banner
+
+Secondary entry point of `@kirbydesign/extensions-angular`. It can be used by importing from `@kirbydesign/extensions-angular/banner`.
