+++ libs/extensions/angular/banner/ng-package.json
