+++ libs/extensions/angular/banner/src/banner.component.html
