+++ libs/extensions/angular/banner/src/banner.component.spec.ts
