+++ libs/extensions/angular/banner/src/banner.component.ts
