+++ libs/extensions/angular/banner/src/index.ts
