+++ libs/extensions/angular/index.ts
@@ -1,0 +1,1 @@
+export * from '@kirbydesign/extensions-angular/banner';
