+++ libs/extensions/angular/ng-package.json
