+++ libs/extensions/angular/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@kirbydesign/composed-ui-angular",
+  "name": "@kirbydesign/extensions-angular",
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^17.0.0",
