+++ libs/extensions/angular/project.json
@@ -1,7 +1,7 @@
 {
-  "name": "composed-ui-angular",
+  "name": "extensions-angular",
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/composed-ui/angular",
+  "sourceRoot": "libs/extensions/angular",
   "prefix": "kirby-x",
   "tags": [],
   "projectType": "library",
@@ -10,14 +10,14 @@
       "executor": "@nx/angular:package",
       "outputs": ["{workspaceRoot}/dist/{projectRoot}"],
       "options": {
-        "project": "libs/composed-ui/angular/ng-package.json"
+        "project": "libs/extensions/angular/ng-package.json"
       },
       "configurations": {
         "production": {
-          "tsConfig": "libs/composed-ui/angular/tsconfig.lib.prod.json"
+          "tsConfig": "libs/extensions/angular/tsconfig.lib.prod.json"
         },
         "development": {
-          "tsConfig": "libs/composed-ui/angular/tsconfig.lib.json"
+          "tsConfig": "libs/extensions/angular/tsconfig.lib.json"
         }
       },
       "defaultConfiguration": "production"
@@ -26,7 +26,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/composed-ui/angular/jest.config.ts"
+        "jestConfig": "libs/extensions/angular/jest.config.ts"
       }
     }
   }
