+++ libs/extensions/angular/tsconfig.json
