+++ libs/extensions/angular/tsconfig.lib.prod.json
